Some kind of ready

+++ src/include/read_obj.h
@@ -8,7 +8,7 @@ class read_obj:basereader
 public:
     read_obj();
     read_obj(const string fileStr);
-    int getGeom(vector<Eigen::Vector3d>& vertE,vector<array<int, 3>>& facE);
+    int getGeom(vector<Eigen::Vector3d>& vertE,vector<array<size_t, 3>>& facE);
 };
 
 #endif // READ_OBJ_H
